+++ MID/MID/Presentation/MonthTodo/Views/Components/MonthTodoCalendarCollectionViewCell.swift
@@ -22,6 +22,7 @@ final class CalendarCollectionViewCell: UICollectionViewCell {
     override init(frame: CGRect) {
         super.init(frame: frame)
         self.configure()
+        backgroundColor = .gray600
     }
     
 //    override func prepareForReuse() {
@@ -35,6 +36,7 @@ final class CalendarCollectionViewCell: UICollectionViewCell {
     private func configure() {
         self.addSubview(self.dayLabel)
         self.dayLabel.font = .boldSystemFont(ofSize: 12)
+        self.dayLabel.textColor = .white000
         self.dayLabel.translatesAutoresizingMaskIntoConstraints = false
         NSLayoutConstraint.activate([
             self.dayLabel.topAnchor.constraint(equalTo: self.topAnchor, constant: 5),
